fix(playwright): run the quarantine drift guard in CI, where it belongs
Two review findings, both correct.

The corpus guardrail keeps playwright/ TypeScript-only precisely because
Playwright's default testMatch collects .mjs — quarantine-list.check.mjs sat
there and would have been picked up as an unlinted test. That broke UI
Checkstyle, and the fix is the same move that answers the second finding: the
guard is now a node:test at playwright/eslint-rules/tests/, a directory that is
exempt from the corpus rule AND already globbed by `yarn test:eslint-rules`,
which UI Checkstyle already runs. It went from "runs when a human remembers" to
running on every PR. Verified both ways: 115/115 pass, and renaming an entry
fails the job with `MATCHES NOTHING`.

The config comment also claimed quarantining "costs the author nothing", which
is only true at retries >= 1 on pull_request. #32833 proposes retries: 0 there,
which inverts the trade. The comment now states the dependency instead of
asserting the conclusion, and names what would have to change.
